Fixing interface calls that have CallData Location (#672)
* Changing data location source from call to definition

* Adding test

* Pushing private function correction

* Fixing test format
